dystematic/requirements.txt
Django==3.0.6
librabbitmq==2.0.0
celery==4.4.2
psycopg2-binary==2.8.5
gunicorn==20.0.4
yfinance==0.1.54
lxml==4.5.1
ipdb==0.13.2
djangorestframework==3.11.0
django-filter==2.2.0